requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
sshtunnel>=0.4.0
paramiko>=2.7.2,<4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0